Fix: All amount commands int type change to float.

+++ swap/cli/providers/bitcoin/fund.py
@@ -14,7 +14,7 @@ from ....providers.config import bitcoin as config
 @click.option("-a", "--address", type=str, required=True, help="Set Bitcoin sender address.")
 @click.option("-ha", "--htlc-address", type=str, required=True,
               help="Set Bitcoin Hash Time Lock Contract (HTLC) address.")
-@click.option("-am", "--amount", type=int, required=True, help="Set Bitcoin fund amount.")
+@click.option("-am", "--amount", type=float, required=True, help="Set Bitcoin fund amount.")
 @click.option("-u", "--unit", type=str, default=config["unit"],
               help="Set Bitcoin fund amount unit.", show_default=True)
 @click.option("-n", "--network", type=str, default=config["network"],
